fix: Windows build - cast VK_CONTROL/VK_V to u8 for keybd_event

+++ src-tauri/src/output/keyboard.rs
@@ -100,10 +100,10 @@ impl KeyboardOutput {
                     keybd_event, KEYEVENTF_KEYUP, VK_CONTROL, VK_V,
                 };
                 unsafe {
-                    keybd_event(VK_CONTROL, 0, 0, 0);
-                    keybd_event(VK_V, 0, 0, 0);
-                    keybd_event(VK_V, 0, KEYEVENTF_KEYUP, 0);
-                    keybd_event(VK_CONTROL, 0, KEYEVENTF_KEYUP, 0);
+                    keybd_event(VK_CONTROL as u8, 0, 0, 0);
+                    keybd_event(VK_V as u8, 0, 0, 0);
+                    keybd_event(VK_V as u8, 0, KEYEVENTF_KEYUP, 0);
+                    keybd_event(VK_CONTROL as u8, 0, KEYEVENTF_KEYUP, 0);
                 }
             }
